+++ Assets/ResonanceAudio/Editor/ResonanceAudioReverbBakingWindow.cs
@@ -25,26 +25,23 @@ using System.Collections.Generic;
 public class ResonanceAudioReverbBakingWindow : EditorWindow {
   private SerializedProperty reverbLayerMask = null;
   private SerializedProperty includeNonStaticGameObjects = null;
-  private SerializedProperty materialMappingGuids = null;
-  private SerializedProperty materialMappingSurfaceMaterials = null;
+  private SerializedProperty materialMap = null;
 
-  private GUIContent reverbProbesBakingLabel = new GUIContent("Bake Reverb To Probes",
+  private GUIContent materialMapLabel = new GUIContent("Material Map",
+      "ResonanceAudioMaterialMap asset to use.");
+  private GUIContent reverbLayerMaskLabel = new GUIContent("Reverb Mask",
+      "Which layers of game objects are included in reverb computation.");
+  private GUIContent nonStaticGameObjectLabel = new GUIContent("Include Non-Static Game Objects",
+      "Should non-static game objects be included in reverb computation?");
+  private GUIContent visualizeModeLabel = new GUIContent("Visualize Mode",
+      "Toggle to visualize the material mapping in the Scene View.");
+  private GUIContent selectReverbProbesLabel = new GUIContent("Select Reverb Probes",
       "Reverb probe selections for baking.");
   private GUIContent selectAllProbesLabel = new GUIContent("Select All",
       "Selects all reverb probes.");
   private GUIContent clearAllProbesLabel = new GUIContent("Clear",
       "Clears reverb probe selections.");
   private GUIContent bakeLabel = new GUIContent("Bake", "Bake reverb to selected reverb probes.");
-  private GUIContent materialMappingLabel = new GUIContent("Map Materials",
-      "Maps Unity Materials or Terrains to surface materials.");
-  private GUIContent visualizeModeLabel = new GUIContent("Visualize Mode",
-      "Toggle to visualize the material mapping in the Scene View.");
-  private GUIContent reverbLayerMaskLabel = new GUIContent("Reverb Mask",
-      "Which layers of game objects are included in reverb computation.");
-  private GUIContent nonStaticGameObjectLabel = new GUIContent("Include Non-Static Game Objects",
-      "Should non-static game objects be included in reverb computation?");
-  private GUIContent clearAllMappingLabel = new GUIContent("Reset All",
-      "Resets the material mapping selections to default.");
 
   // Whether to visualize the material mapping.
   private bool isInVisualizeMode = false;
@@ -67,63 +64,21 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
   // not be available yet).
   private HashSet <int> updatedSceneViews = null;
 
-  // The thumbnail previews of the surface materials in the inspector window, shown as solid color
-  // patches.
-  private Texture2D[] surfaceMaterialPreviews = null;
-
   // Shader to visualize surface materials.
   private Shader surfaceMaterialShader = null;
 
-  private const int guidLabelWidth = 150;
-  private const int materialRowMargin = 5;
-  private const int materialPreviewSize = 50;
-  private GUILayoutOption previewHeight = null;
-  private GUILayoutOption previewWidth = null;
-  private GUIStyle materialRowStyle = null;
+  // This is to accomodate the long |nonStaticGameObjectLabel|.
+  private const float propertyLabelWidth = 184.0f;
 
-  // The scroll position.
-  private Vector2 scrollPosition = new Vector2();
+  // The scroll position of the reverb probe selection UI.
+  private Vector2 probeSelectionScrollPosition = Vector2.zero;
 
-  // The foldouts of the reverb baking and material mapping sections.
-  private bool showReverbBaking = true;
-  private bool showMaterialMapping = true;
+  // The foldout of the reverb probe selection.
+  private bool showReverbProbeSelection = true;
 
   // The path to the material mapper asset.
   private const string materialMapperAssetPath =
       "Assets/ResonanceAudio/Resources/ResonanceAudioMaterialMapper.asset";
-
-  // Color coding for surface materials used to visualize the surface material assignment in the
-  // scene as well as in the material picking UI.
-  // The following colors are generated using http://vrl.cs.brown.edu/color, while setting a
-  // starting point of "rgb(128,128,128)", and maximizing the "Perceptual Distance" and
-  // "Pair Preference" parameters. This color mapping is shared by the shader and the preview
-  // thumbnails of the surface materials (see InitializeColorArrayInShader() and
-  // InitializeSurfaceMaterialPreviews()).
-  private static readonly Color[] surfaceMaterialColors = new Color[] {
-      new Color(0.500000f, 0.500000f, 0.500000f),
-      new Color(0.545098f, 0.909804f, 0.678431f),
-      new Color(0.184314f, 0.258824f, 0.521569f),
-      new Color(0.552941f, 0.737255f, 0.976471f),
-      new Color(0.035294f, 0.376471f, 0.074510f),
-      new Color(0.952941f, 0.415686f, 0.835294f),
-      new Color(0.105882f, 0.894118f, 0.427451f),
-      new Color(0.541176f, 0.015686f, 0.345098f),
-      new Color(0.631373f, 0.847059f, 0.196078f),
-      new Color(0.513725f, 0.003922f, 0.741176f),
-      new Color(0.949020f, 0.690196f, 0.964706f),
-      new Color(0.082353f, 0.305882f, 0.337255f),
-      new Color(0.152941f, 0.792157f, 0.901961f),
-      new Color(0.921569f, 0.070588f, 0.254902f),
-      new Color(0.274510f, 0.635294f, 0.423529f),
-      new Color(0.556863f, 0.215686f, 0.066667f),
-      new Color(0.960784f, 0.803922f, 0.686275f),
-      new Color(0.305882f, 0.282353f, 0.035294f),
-      new Color(0.917647f, 0.839216f, 0.141176f),
-      new Color(0.521569f, 0.458824f, 0.858824f),
-      new Color(0.937255f, 0.592157f, 0.176471f),
-      new Color(0.980392f, 0.105882f, 0.988235f),
-      new Color(0.725490f, 0.423529f, 0.552941f)
-  };
 
   [MenuItem("ResonanceAudio/Reverb Baking")]
   private static void Initialize() {
@@ -136,13 +91,10 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     updatedSceneViews = new HashSet<int>();
 
     InitializeColorArrayInShader();
-    InitializeGuiParameters();
-    InitializeSurfaceMaterialPreviews();
     InitializeSurfaceMaterialShader();
 
     LoadOrCreateMaterialMapper();
     LoadOrCreateMaterialMapperUpdater();
-    InitializeMaterialMapperAndLoadProperties();
 
     isInVisualizeMode = false;
 
@@ -183,34 +135,35 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     serializedMaterialMapper.Update();
 
     EditorGUI.BeginDisabledGroup(EditorApplication.isPlaying);
-    showReverbBaking = EditorGUILayout.Foldout(showReverbBaking, reverbProbesBakingLabel);
-    if (showReverbBaking) {
+    var savedWidth = EditorGUIUtility.labelWidth;
+    EditorGUIUtility.labelWidth = propertyLabelWidth;
+    DrawMaterialMapSelection();
+
+    EditorGUILayout.Separator();
+
+    EditorGUI.BeginDisabledGroup(materialMap.objectReferenceValue == null);
+    DrawObjectFiltering();
+
+    EditorGUILayout.Separator();
+
+    DrawVisualizeModeCheckbox();
+    EditorGUIUtility.labelWidth = savedWidth;
+
+    EditorGUILayout.Separator();
+
+    showReverbProbeSelection = EditorGUILayout.Foldout(showReverbProbeSelection,
+                                                       selectReverbProbesLabel);
+    if (showReverbProbeSelection) {
       ++EditorGUI.indentLevel;
       DrawProbeSelection();
-
-      EditorGUILayout.Separator();
-
-      DrawBakeButton();
       --EditorGUI.indentLevel;
     }
 
     EditorGUILayout.Separator();
 
-    showMaterialMapping = EditorGUILayout.Foldout(showMaterialMapping, materialMappingLabel);
-    if (showMaterialMapping) {
-      ++EditorGUI.indentLevel;
-      DrawVisualizeModeCheckbox();
-
-      EditorGUILayout.Separator();
-
-      DrawObjectFiltering();
-
-      EditorGUILayout.Separator();
-
-      DrawMaterialMappingGUI();
-      --EditorGUI.indentLevel;
-    }
-    EditorGUI.EndDisabledGroup();
+    DrawBakeButton();
+    EditorGUI.EndDisabledGroup();  // Disabled if no material map is selected.
+    EditorGUI.EndDisabledGroup();  // Disabled if in Play mode.
 
     serializedMaterialMapper.ApplyModifiedProperties();
   }
@@ -227,6 +180,14 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     }
 
     serializedMaterialMapper = new UnityEditor.SerializedObject(materialMapper);
+    reverbLayerMask = serializedMaterialMapper.FindProperty("reverbLayerMask");
+    includeNonStaticGameObjects =
+        serializedMaterialMapper.FindProperty("includeNonStaticGameObjects");
+    materialMap = serializedMaterialMapper.FindProperty("materialMap");
+
+    materialMapper.Initialize();
+    RefreshMaterialMapper();
+    UpdateShader();
   }
 
   // Loads the unique material mapper updater; creates one if not found.
@@ -264,22 +225,7 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
 
       materialMapperUpdater = newUpdater;
     }
-    materialMapperUpdater.RefreshMaterialMapper = RefreshMaterialMapper;
-  }
-
-  // Initializes the material mapper and loads properties to be displayed in this window.
-  private void InitializeMaterialMapperAndLoadProperties() {
-    materialMapper.Initialize();
-    RefreshMaterialMapper();
-    UpdateShader();
-
-    reverbLayerMask = serializedMaterialMapper.FindProperty("reverbLayerMask");
-    includeNonStaticGameObjects =
-        serializedMaterialMapper.FindProperty("includeNonStaticGameObjects");
-    var surfaceMaterialFromGuid = serializedMaterialMapper.FindProperty("surfaceMaterialFromGuid");
-    materialMappingGuids = surfaceMaterialFromGuid.FindPropertyRelative("guids");
-    materialMappingSurfaceMaterials =
-        surfaceMaterialFromGuid.FindPropertyRelative("surfaceMaterials");
+    materialMapperUpdater.RefreshMaterialMapper = RefreshMaterialMapperOnlyInVisualizeMode;
   }
 
   // Initializes the surface material colors in a global vector array for shaders.
@@ -289,40 +235,11 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     Vector4[] vectorArray = new Vector4[numSurfaceMaterials];
     for (int surfaceMaterialIndex = 0; surfaceMaterialIndex < numSurfaceMaterials;
          ++surfaceMaterialIndex) {
-      var color = surfaceMaterialColors[surfaceMaterialIndex];
+      var color = ResonanceAudioMaterialMap.surfaceMaterialColors[surfaceMaterialIndex];
       vectorArray[surfaceMaterialIndex] = new Vector4(color.r, color.g, color.b, 0.5f);
     }
 
     Shader.SetGlobalVectorArray("_SurfaceMaterialColors", vectorArray);
-  }
-
-  // Initializes various GUI parameters.
-  private void InitializeGuiParameters() {
-    previewHeight = GUILayout.Height((float) materialPreviewSize);
-    previewWidth = GUILayout.Width((float) materialPreviewSize);
-    materialRowStyle = new GUIStyle();
-    materialRowStyle.margin = new RectOffset(materialRowMargin, materialRowMargin,
-                                             materialRowMargin, materialRowMargin);
-  }
-
-  // Initializes the thumbnail previews used in the material picking UI. Each surface material
-  // is shown as a square filled with solid color.
-  private void InitializeSurfaceMaterialPreviews () {
-    int numSurfaceMaterials = surfaceMaterialColors.Length;
-    surfaceMaterialPreviews = new Texture2D[numSurfaceMaterials];
-    for (int surfaceMaterialIndex = 0; surfaceMaterialIndex < numSurfaceMaterials;
-         ++surfaceMaterialIndex) {
-      var color = surfaceMaterialColors[surfaceMaterialIndex];
-      Texture2D surfaceMaterialPreview = new Texture2D(materialPreviewSize, materialPreviewSize);
-      var pixelArraySize = surfaceMaterialPreview.GetPixels().Length;
-      Color[] pixelArray = new Color[pixelArraySize];
-      for (int pixelArrayIndex = 0; pixelArrayIndex < pixelArraySize; ++pixelArrayIndex) {
-        pixelArray[pixelArrayIndex] = color;
-      }
-      surfaceMaterialPreview.SetPixels(pixelArray);
-      surfaceMaterialPreview.Apply();
-      surfaceMaterialPreviews[surfaceMaterialIndex] = surfaceMaterialPreview;
-    }
   }
 
   // Initializes the surface material shader which visualizes surface materials as colors.
@@ -341,6 +258,10 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
       return;
     }
 
+    if (materialMap.objectReferenceValue == null) {
+      return;
+    }
+
     MeshRenderer[] meshRenderers = null;
     List<string>[] guidsForMeshRenderers = null;
     GatherMeshRenderersAndGuids(ref meshRenderers, ref guidsForMeshRenderers);
@@ -351,6 +272,13 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     materialMapper.ApplyMaterialMapping(meshRenderers, guidsForMeshRenderers, activeTerrains,
                                         guidsForTerrains, surfaceMaterialShader);
     redraw = true;
+  }
+
+  // Refreshes the material mapper's data only in visualize mode.
+  private void RefreshMaterialMapperOnlyInVisualizeMode() {
+    if (isInVisualizeMode) {
+      RefreshMaterialMapper();
+    }
   }
 
   // Gathers the mesh renderes of game objects, and the GUIDs of the Unity Materials of
@@ -440,6 +368,30 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     updatedSceneViews.Add(sceneView.GetInstanceID());
   }
 
+  // The UI for selecting a ResonanceAudioMaterialMap asset to use.
+  private void DrawMaterialMapSelection() {
+    EditorGUILayout.PropertyField(materialMap, materialMapLabel);
+  }
+
+  // Draws the objects filtering GUI. Users can decide which layers to include, and whether to
+  // include non-static objects.
+  private void DrawObjectFiltering() {
+    EditorGUILayout.PropertyField(reverbLayerMask, reverbLayerMaskLabel);
+
+    EditorGUILayout.Separator();
+
+    EditorGUILayout.PropertyField(includeNonStaticGameObjects, nonStaticGameObjectLabel);
+  }
+
+  // Draws the "Visualize Mode" checkbox.
+  private void DrawVisualizeModeCheckbox() {
+    if (isInVisualizeMode != EditorGUILayout.Toggle(visualizeModeLabel, isInVisualizeMode)) {
+      isInVisualizeMode = !isInVisualizeMode;
+      RefreshMaterialMapper();
+      UpdateShader();
+    }
+  }
+
   // The UI for selecting a subset of reverb probes to bake reverb to.
   private void DrawProbeSelection() {
     ResonanceAudioReverbProbe[] allReverbProbes =
@@ -449,6 +401,8 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     var selectedReverbProbes = ResonanceAudioReverbComputer.selectedReverbProbes;
     selectedReverbProbes.RemoveAll(reverbProbe => reverbProbe == null);
 
+    probeSelectionScrollPosition = EditorGUILayout.BeginScrollView(probeSelectionScrollPosition,
+                                                                   GUILayout.ExpandHeight(false));
     for (int i = 0; i < allReverbProbes.Length; ++i) {
       var reverbProbe = allReverbProbes[i];
       bool currentlySelected = selectedReverbProbes.Contains(reverbProbe);
@@ -464,6 +418,8 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
         }
       }
     }
+    EditorGUILayout.EndScrollView();
+
     if (allReverbProbes.Length > 0) {
       EditorGUILayout.Separator();
 
@@ -508,101 +464,6 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
     EditorGUI.EndDisabledGroup();
   }
 
-  // Draws the "Visualize Mode" checkbox.
-  private void DrawVisualizeModeCheckbox() {
-    if (isInVisualizeMode != EditorGUILayout.Toggle(visualizeModeLabel, isInVisualizeMode)) {
-      isInVisualizeMode = !isInVisualizeMode;
-      RefreshMaterialMapper();
-      UpdateShader();
-    }
-  }
-
-  // Draws the objects filtering GUI. Users can decide which layers to include, and whether to
-  // include non-static objects.
-  private void DrawObjectFiltering() {
-    EditorGUILayout.PropertyField(reverbLayerMask, reverbLayerMaskLabel);
-
-    EditorGUILayout.Separator();
-
-    EditorGUILayout.PropertyField(includeNonStaticGameObjects, nonStaticGameObjectLabel);
-  }
-
-  // Draws the material mapping GUI. The GUI is organized as rows, each row having a GUID (a Unity
-  // Material or a terrain data) on the left, and the mapped surface materials on the right.
-  private void DrawMaterialMappingGUI() {
-    // Show the material mapping as rows.
-    scrollPosition = EditorGUILayout.BeginScrollView(scrollPosition, GUILayout.ExpandHeight(false));
-    // Access the SurfaceMaterialDictionary's underlying two serialized lists directly.
-    for (int i = 0; i < materialMappingGuids.arraySize; ++i) {
-      EditorGUILayout.BeginHorizontal(materialRowStyle);
-      GUILayout.Space(15 * EditorGUI.indentLevel);
-      if (DrawGuidColumn(materialMappingGuids.GetArrayElementAtIndex(i).stringValue)) {
-        DrawSurfaceMaterialColumn(materialMappingSurfaceMaterials.GetArrayElementAtIndex(i));
-      }
-      EditorGUILayout.EndHorizontal();
-    }
-    EditorGUILayout.EndScrollView();
-
-    EditorGUILayout.Separator();
-
-    DrawClearAllButton();
-  }
-
-  // Draws the GUID column: the thumbnail preview first, followed by the name. Depending on whether
-  // the GUID identifies a Unity Material or a terrain data, shows the previews and names
-  // differently.
-  // Returns false if the GUID does not correspond to a Unity Material or a terrain data (maybe
-  // not present in the loaded scenes).
-  private bool DrawGuidColumn(string guid) {
-    // Select the preview and the name.
-    Texture2D assetPreview = null;
-    string guidName = null;
-    var unityMaterial = materialMapper.GetUnityMaterial(guid);
-    var terrainData = materialMapper.GetTerrainData(guid);
-    if (unityMaterial != null) {
-      assetPreview = AssetPreview.GetAssetPreview(unityMaterial);
-      guidName = unityMaterial.name;
-    } else if (terrainData != null) {
-      assetPreview = AssetPreview.GetMiniThumbnail(terrainData);
-      guidName = terrainData.name;
-    } else {
-      // Both |unityMaterial| and |terrainData| are NULL; display nothing for this GUID.
-      return false;
-    }
-
-    // Draw the preview.
-    GUILayout.Box(assetPreview, GUIStyle.none, previewHeight, previewWidth);
-
-    // Display the name.
-    if (guidName != null) {
-      EditorGUILayout.LabelField(guidName, GUILayout.Width(guidLabelWidth));
-    }
-    return true;
-  }
-
-  // Draws the surface material column: the thumbnail preview first, followed by a drop-down menu
-  // to let users choose the mapped material.
-  private void DrawSurfaceMaterialColumn(SerializedProperty surfaceMaterialProperty) {
-    // Draw the preview.
-    var preview = surfaceMaterialPreviews[surfaceMaterialProperty.enumValueIndex];
-    GUILayout.Box(preview, GUIStyle.none, previewHeight, previewWidth);
-
-    // Draw the drop-down menu.
-    EditorGUILayout.PropertyField(surfaceMaterialProperty, GUIContent.none);
-  }
-
-  // Draws the "Clear All" button and clears the material mapping (by clearing the underlying
-  // serialized lists).
-  private void DrawClearAllButton() {
-    EditorGUILayout.BeginHorizontal();
-    GUILayout.Space(15 * EditorGUI.indentLevel);
-    if (GUILayout.Button(clearAllMappingLabel)) {
-      materialMappingGuids.ClearArray();
-      materialMappingSurfaceMaterials.ClearArray();
-    }
-    EditorGUILayout.EndHorizontal();
-  }
-
   private void OnSceneGUI(SceneView sceneView) {
     // Deferred update of the scene view if it is not updated yet.
     if (!updatedSceneViews.Contains(sceneView.GetInstanceID())) {
@@ -617,9 +478,6 @@ public class ResonanceAudioReverbBakingWindow : EditorWindow {
 
   private void OnSceneOrModeSwitch() {
     LoadOrCreateMaterialMapperUpdater();
-
-    // Switching scenes play modes destroys the color-coding Texture2D. Re-initialize them.
-    InitializeSurfaceMaterialPreviews();
 
     // Force repaint this window to reflect the scene changes, which may have a different set of
     // Unity Materials and Terrain data.