+++ RELEASE_NOTES.md
@@ -1,11 +1,29 @@
 # Release notes
 
-## Resonance Audio SDK for Unity v1.1.1
+## Resonance Audio SDK for Unity v1.2.0
+
+### Additions
+* Added advanced near-field effect for sound sources less than 1 meter from the listener. Introduced `ResonanceAudioSource.nearFieldEffectEnabled` and `ResonanceAudioSource.nearFieldEffectGain` parameters to simulate the effect of sound sources being very close to the listener's ears. Note that this effect could result in up to ~9x gain boost on the source input. Therefore, it is advised to set smaller gain values or reduce the input gain for louder sound sources to avoid clipping of the output signal.
+* Added `ResonanceAudioSource.occlusionIntensity` parameter to adjust the intensity of the occlusion effect.
+
+### Behavioral Changes
+* Significant CPU performance improvement for reverb times more than 0.6 seconds (thanks to a new spectral reverb implementation under the hood). Also, delivers a slightly brighter sounding reverb.
+* Increased `ResonanceAudio.maxReverbTime` to 10 seconds for the `ResonanceAudioRoom.reverbTime` and `ResonanceAudioReverbProbe.reverbTime` modifiers.
+* Material maps are stored as assets and are modified in the Inspector Window instead of the `Reverb Baking` window. Multiple maps per project are allowed, and the Reverb Baking Window selects one to be used in reverb computation. If you are updating the SDK from an older version, please make sure to remove the `ResonanceAudio` folder from the project assets before importing the new SDK to avoid potential asset conflicts.
+* Improved the reverb gain adjustment parameter in `ResonanceAudioRoom` and `ResonanceAudioReverbProbe` to avoid noticeable delays with larger gain changes.
+
+### Bug fixes
+* Addressed [issue #16](https://github.com/resonance-audio/resonance-audio-unity-sdk/issues/16) where the `Reverb Baking` window would constantly lock up the editor in a project with a lot of materials. Note that there may still be a significant performance overhead if the `Visualize Mode` is enabled while making changes in the scene hierarchy, such as transform changes of a game object.
+* Fixed [issue #11](https://github.com/resonance-audio/resonance-audio-unity-sdk/issues/11) where updating SDK would override material mappings.
+* Fixed [issue #13](https://github.com/resonance-audio/resonance-audio-unity-sdk/issues/13) where the list of 'ResonanceAudioReverbProbe' was not scrollable in Reverb Baking Window.
+* Fixed a bug where the ResonanceAudioAcousticMesh class would give an error when the scene contains non-triangular meshes (e.g. points and lines). Now non-triangular meshes will be skipped.
+
+## Resonance Audio SDK for Unity v1.1.1 (2017-12-18)
 
 ### Bug fixes
 * Fixed a bug in stereo deinterleaving input buffer conversion that could lead to a crash.
 
-## Resonance Audio SDK for Unity v1.1.0
+## Resonance Audio SDK for Unity v1.1.0 (2017-12-15)
 
 ### Additions
 * Added multi-object editing support to `ResonanceAudioReverbProbe`.
@@ -21,6 +39,7 @@
 * Fixed [issue #4](https://github.com/resonance-audio/resonance-audio-unity-sdk/issues/4) where the enable callback of each `ResonanceAudioSource` component would cause GC Allocs at run time.
 * Fixed [issue #8](https://github.com/resonance-audio/resonance-audio-unity-sdk/issues/8) where the Unity Editor could occasionally crash when the `ResonanceAudioReverbProbe` component is used in the scene.
 * Fixed an issue where the `ResonanceAudioListener` would unnecessarily make update calls in Edit Mode.
+* Fixed a depracation warning in `ResonanceAudioReverbBakingWindow` in Unity 2017.2+.
 
 ### Other Changes
 * For projects upgrading from Google VR Audio to Resonance Audio, the SDK will now detect incompatible assets at build time and offer to remove them. See additional [upgrade instructions](https://developers.google.com/resonance-audio/migrate/).
@@ -37,5 +56,4 @@ This is the initial release of Resonance Audio SDK for Unity, which includes:
 * Sound directivity controls
 * Source spread controls
 * Distance attenuation
-
 * Near-field rendering